requirements.txt
firebase-admin==6.2.0
google-cloud-language==2.10.1
python-dotenv==1.0.0